feat(import): merge service import into the file import dialog
Collapse the separate "Import files" and "Import from a service" dialogs
into one. The unified dialog keeps the service picker cards (services in
a 3-per-row grid) and adds local file import as a full-width "Files &
other formats" card grouped beneath them, selected by default.

- Rename widgets/dialogs/import_providers/ -> import/ and move both the
  dialog container (import_provider.tsx -> import/import_dialog.tsx) and
  the old files dialog (import.tsx -> import/files.tsx, now a provider
  panel) into it.
- The merged dialog listens to the old showImportDialog event; the
  showImportProviderDialog event, its LazyDialog registration and the
  duplicate "Import from a service" note-actions entry are removed.
- ImportProvider gains optional `icon` (Boxicons) and `group`
  ("service"|"local") so the Files card can render bx-import and sit in
  its own full-width group.
