+++ apps/client/src/widgets/dialogs/import/types.ts
@@ -18,10 +18,14 @@ export interface ImportProvider {
     id: string;
     /** Human-readable name shown in the provider picker. */
     name: string;
-    /** URL of the provider's logo (an imported SVG, e.g. `import iconUrl from "./icons/notion.svg?url"`). Rendered monochrome via a CSS mask so it adapts to the theme. */
-    iconUrl: string;
+    /** URL of the provider's logo (an imported SVG, e.g. `import iconUrl from "./icons/notion.svg?url"`). Rendered monochrome via a CSS mask so it adapts to the theme. Provide either this or {@link icon}. */
+    iconUrl?: string;
+    /** Boxicons class (e.g. `"bx bx-import"`) shown in place of {@link iconUrl}. Provide either this or `iconUrl`. */
+    icon?: string;
     /** One-line description shown under the name. */
     description: string;
+    /** Visual grouping in the picker: `"service"` (the default) appears in the card grid; `"local"` (file upload) is shown full-width, grouped beneath the services and selected by default. */
+    group?: "service" | "local";
     /** The component rendered once this provider is chosen; it drives its own multi-step flow. */
     Panel: ComponentType<ImportProviderPanelProps>;
 }
