--- apps/client/src/widgets/dialogs/import_providers/index.ts
@@ -1,14 +1,0 @@
-import anytypeProvider from "./anytype.js";
-import evernoteProvider from "./evernote.js";
-import keepProvider from "./keep.js";
-import notionProvider from "./notion.js";
-import oneNoteProvider from "./onenote.js";
-import type { ImportProvider } from "./types.js";
-
-/**
- * Registry of available import providers. Append new providers here; the generic import dialog
- * renders the picker and each provider's panel automatically.
- */
-export const importProviders: ImportProvider[] = [oneNoteProvider, notionProvider, keepProvider, evernoteProvider, anytypeProvider];
-
-export type { ImportProvider, ImportProviderPanelProps } from "./types.js";
